requirements.txt
# requirements.txt
openai>=1.0.0
pdfplumber>=0.10.0
numpy>=1.21.0
pyyaml>=6.0.0
python-dotenv>=0.19.0
logging>=0.5.0
aiohttp>=3.8.0